pyproject.toml
[tool.poetry]
name = "delivery_tracking_etl"
version = "0.1.0"
description = "Data Importer ETL"
authors = ["Viko Bastidas <vico.bastidas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "9.1.0"
python-dotenv = "^1.0.1"
pytz = "^2025.1"

[tool.poetry.scripts]
etl_data_importer = "delivery_tracking_etl.etl_data_importer:main"
etl_assign_estimated_delivery_date = "delivery_tracking_etl.etl_assign_estimated_delivery_date:main"
etl_notification_deliverycompleted = "delivery_tracking_etl.etl_notification_deliverycompleted:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
